dev feature: mongodb integration: dev storage methods,fist test passed

+++ aggregation.mysql/src/test/java/com/alibaba/smart/framework/engine/test/api/service/TaskServiceTest.java
@@ -96,7 +96,7 @@ public class TaskServiceTest {
         submitTaskInstanceList=  taskQueryService.findPendingTaskList(pendingTaskQueryParam);
         Assert.assertEquals(1,submitTaskInstanceList.size());
 
-        Integer count =  taskQueryService.countPendingTaskList(pendingTaskQueryParam);
+        Long count =  taskQueryService.countPendingTaskList(pendingTaskQueryParam);
         Assert.assertEquals(1L,count.longValue());
 
         PendingTaskQueryParam  pendingTaskQueryParam1 = new PendingTaskQueryParam();
