+++ aggregation.mysql/src/test/java/com/alibaba/smart/framework/engine/test/process/DefaultMultiInstanceCounter.java
@@ -2,14 +2,9 @@ package com.alibaba.smart.framework.engine.test.process;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
 import com.alibaba.smart.framework.engine.SmartEngine;
 import com.alibaba.smart.framework.engine.configuration.MultiInstanceCounter;
-import com.alibaba.smart.framework.engine.configuration.TaskAssigneeDispatcher;
-import com.alibaba.smart.framework.engine.constant.AssigneeTypeConstant;
-import com.alibaba.smart.framework.engine.model.assembly.Activity;
-import com.alibaba.smart.framework.engine.model.instance.TaskAssigneeCandidateInstance;
 import com.alibaba.smart.framework.engine.service.param.query.TaskInstanceQueryParam;
 import com.alibaba.smart.framework.engine.service.query.TaskQueryService;
 
@@ -19,8 +14,8 @@ import com.alibaba.smart.framework.engine.service.query.TaskQueryService;
 public class DefaultMultiInstanceCounter implements MultiInstanceCounter {
 
 
-    public Integer countPassedTaskInstanceNumber(String processInstanceId, String activityInstanceId,
-                                                 SmartEngine smartEngine) {
+    public Long countPassedTaskInstanceNumber(String processInstanceId, String activityInstanceId,
+                                              SmartEngine smartEngine) {
       TaskQueryService taskQueryService = smartEngine.getTaskQueryService();
 
         TaskInstanceQueryParam taskInstanceQueryParam = new TaskInstanceQueryParam();
@@ -29,14 +24,14 @@ public class DefaultMultiInstanceCounter implements MultiInstanceCounter {
         taskInstanceQueryParam.setProcessInstanceIdList(processInstanceIdList);
         taskInstanceQueryParam.setActivityInstanceId(activityInstanceId);
         taskInstanceQueryParam.setTag(FullMultiInstanceTest.AGREE);
-        Integer count  = taskQueryService.count(taskInstanceQueryParam);
+        Long count  = taskQueryService.count(taskInstanceQueryParam);
         return  count;
 
     }
 
     @Override
-    public Integer countRejectedTaskInstanceNumber(String processInstanceId, String activityInstanceId,
-                                                   SmartEngine smartEngine) {
+    public Long countRejectedTaskInstanceNumber(String processInstanceId, String activityInstanceId,
+                                                SmartEngine smartEngine) {
         TaskQueryService taskQueryService = smartEngine.getTaskQueryService();
 
         TaskInstanceQueryParam taskInstanceQueryParam = new TaskInstanceQueryParam();
@@ -45,7 +40,7 @@ public class DefaultMultiInstanceCounter implements MultiInstanceCounter {
         taskInstanceQueryParam.setProcessInstanceIdList(processInstanceIdList);
         taskInstanceQueryParam.setActivityInstanceId(activityInstanceId);
         taskInstanceQueryParam.setTag(FullMultiInstanceTest.DISAGREE);
-        Integer count  = taskQueryService.count(taskInstanceQueryParam);
+        Long count  = taskQueryService.count(taskInstanceQueryParam);
         return  count;
 
     }
